split cartography presenter
