+++ src/features/cartography/infrastructure/presentation/components/permanence-map/permanence-map.component.spec.ts
@@ -3,6 +3,11 @@ import { PermanenceMapComponent } from './permanence-map.component';
 import { FailedToCompileError } from '@angular-common/errors';
 import { MARKERS_TOKEN } from '../../../configuration';
 import { LeafletMapStubComponent } from '../../test-doubles';
+import { DepartmentPermanenceMapPresenter } from './department-permanence-map.presenter';
+import { RegionPermanenceMapPresenter } from './region-permanence-map.presenter';
+import { CnfsPermanenceMapPresenter } from './cnfs-permanence-map.presenter';
+import { CartographyPresenter } from '../../pages';
+import { Observable, of } from 'rxjs';
 
 describe('Permanence-map component', (): void => {
   beforeEach(async (): Promise<void> => {
@@ -11,11 +16,39 @@ describe('Permanence-map component', (): void => {
       imports: [],
       providers: [
         {
-          provide: MARKERS_TOKEN,
+          provide: CartographyPresenter,
           useValue: {}
         }
       ]
     })
+      .overrideComponent(PermanenceMapComponent, {
+        set: {
+          providers: [
+            {
+              provide: CnfsPermanenceMapPresenter,
+              useValue: {
+                visibleMapCnfsPermanencesThroughViewportAtZoomLevel$: (): Observable<null> => of(null)
+              }
+            },
+            {
+              provide: DepartmentPermanenceMapPresenter,
+              useValue: {
+                visibleMapCnfsByDepartmentAtZoomLevel$: (): Observable<null> => of(null)
+              }
+            },
+            {
+              provide: RegionPermanenceMapPresenter,
+              useValue: {
+                visibleMapCnfsByRegionAtZoomLevel$: (): Observable<null> => of(null)
+              }
+            },
+            {
+              provide: MARKERS_TOKEN,
+              useValue: {}
+            }
+          ]
+        }
+      })
       .compileComponents()
       .catch((): void => {
         throw new FailedToCompileError('PermanenceMap');
